fix: Use simplified home page to resolve loading issues
- Replace complex home page with simple version
- Remove external video dependencies
- Use basic gradient background
- Keep original as index-full.tsx for later restoration
